+++ libs/ngx-appwrite/changelog.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.5.4] - 2024-03-29
+
+Added messaging service. See https://appwrite.io/docs/references/1.5.x/client-web/messaging
+
+Updated dependencies.
+
 ## [1.5.3] - 2024-03-23
 
 Changed some method names in account service to mirror the official names from the Appwrite Web SDK version 14.0.0